Add initial version of basic filesystem driver
The fs_basic_if interface provides a cut down feature set with the aim
of a low memory footprint.

- To support the seek command in the basic interface the Petit FatFs
  config has been updated to build with seek support

A storage media interface is also defined to allow the filesystems
ported to this library to have a common abstraction layer to the devices
used below them.

- Support for quad SPI flash devices as storage media is also included
  to allow filesystems on the data partition of the boot device via
  libflash


A small test app is included to allow the building of both the basic
filesystem abstraction layer, Petit FatFs, and the QSPI media layer.